% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{vector_1}{A vector to store the first column of the data frame as a vector}

\item{vector_2}{A vector to store the second column of the data frame as a vector}

\item{vector_3}{A vector to store the third column of the data frame as a vector}

\item{unique_values}{To extract the unique values of the first column}
}
\value{
The shortest distances from the input node to all other nodes
}
\description{
Dijkstra Algorithm
}
\examples{
dijkstra(wiki_graph,2)
dijkstra(wiki_graph,1)
}
