% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Greatest Common Divisor[GCD] of two numbers}
\usage{
euclidean(dividend, divisor)
}
\arguments{
\item{dividend}{A number}

\item{divisor}{A number}
}
\value{
The GCD of two numbers
}
\description{
Greatest Common Divisor[GCD] of two numbers
}
\examples{
euclidean(25,2)
euclidean(30,6)
}
