\name{euclidstra-package}
\alias{euclidstra-package}
\alias{euclidstra}
\docType{package}
\title{
\packageTitle{euclidstra}
}
\description{
\packageDescription{euclidstra}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{euclidstra}
\packageIndices{euclidstra}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{euclidstra}

Maintainer: \packageMaintainer{euclidstra}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
